+++ Indexing/Index.hpp
@@ -21,11 +21,9 @@
 #include "Lib/Event.hpp"
 #include "Kernel/Clause.hpp"
 #include "Kernel/Term.hpp"
+#include "Kernel/TypedTermList.hpp"
 #include "Saturation/ClauseContainer.hpp"
 #include "Kernel/Clause.hpp"
-#include "ResultSubstitution.hpp"
-#include "Kernel/UnificationWithAbstraction.hpp"
-#include "Kernel/TermOrderingDiagram.hpp"
 
 /**
  * Indices are parametrized by a LeafData, i.e. the bit of data you want to store in the index.
@@ -110,50 +108,9 @@ struct TermLiteralClause
                << ")"; }
 };
 
-/** Custom leaf data for forward demodulation to store the demodulator
- * left- and right-hand side normalized and cache preorderedness. */
-struct DemodulatorData
-{
-  DemodulatorData(TypedTermList term, TypedTermList rhs, Clause* clause, bool preordered, const Ordering& ord)
-    : term(term), rhs(rhs), clause(clause), preordered(preordered), tod(ord.createTermOrderingDiagram())
-  {
-    // insert pointer to owner as non-null value representing success
-    tod->insert({ { term, rhs, Ordering::GREATER } }, this);
-#if VDEBUG
-    ASS(term.containsAllVariablesOf(rhs));
-    ASS(!preordered || ord.compare(term,rhs)==Ordering::GREATER);
-    Renaming r;
-    r.normalizeVariables(term);
-    ASS_EQ(term,r.apply(term));
-    ASS_EQ(rhs,r.apply(rhs));
-#endif
-  }
-
-  // lhs, the identifier is required to be `term` by CodeTree
-  TypedTermList term;
-  TermList rhs;
-  Clause* clause;
-  bool preordered; // whether term > rhs
-  TermOrderingDiagramUP tod; // TOD for checking term > rhs
-
-  TypedTermList const& key() const { return term; }
-
-  auto asTuple() const
-  { return std::make_tuple(clause->number(), term, rhs); }
-
-  IMPL_COMPARISONS_FROM_TUPLE(DemodulatorData)
-
-  friend std::ostream& operator<<(std::ostream& out, DemodulatorData const& self)
-  { return out << self.term.untyped() << " = " << self.rhs << ", " << self.clause->toString(); }
-};
-
 template<class T>
 struct is_indexed_data_normalized
 { static constexpr bool value = false; };
-
-template<>
-struct is_indexed_data_normalized<DemodulatorData>
-{ static constexpr bool value = true; };
 
 /**
  * Class of objects which contain results of term queries.
