+++ Inferences/BackwardDemodulation.hpp
@@ -17,7 +17,7 @@
 #define __BackwardDemodulation__
 
 #include "Forwards.hpp"
-#include "Indexing/TermIndex.hpp"
+#include "Indexing/DemodulationIndex.hpp"
 
 #include "DemodulationHelper.hpp"
 #include "InferenceEngine.hpp"
