+++ Inferences/Cancellation.cpp
@@ -11,6 +11,7 @@
 #include "Inferences/Cancellation.hpp"
 #include "Kernel/Ordering.hpp"
 #include "Kernel/NumTraits.hpp"
+#include "Kernel/PolynomialBottomUpEvaluation.hpp"
 
 #define DEBUG(...) //DBG(__VA_ARGS__)
