+++ Inferences/ForwardDemodulation.hpp
@@ -17,7 +17,7 @@
 #define __ForwardDemodulation__
 
 #include "Forwards.hpp"
-#include "Indexing/TermIndex.hpp"
+#include "Indexing/DemodulationIndex.hpp"
 
 #include "DemodulationHelper.hpp"
 #include "InferenceEngine.hpp"
